Conflicts:
	src/Sylius/Bundle/ResourceBundle/Behat/DefaultContext.php

+++ src/Sylius/Bundle/ShippingBundle/Form/Type/Calculator/VolumeRateConfigurationType.php
@@ -51,7 +51,7 @@ class VolumeRateConfigurationType extends AbstractType
                 'label' => 'sylius.form.shipping_calculator.volume_rate_configuration.amount',
                 'constraints' => array(
                     new NotBlank(),
-                    new Type(array('type' => 'numeric')),
+                    new Type(array('type' => 'integer')),
                 )
             ))
             ->add('division', 'number', array(
