+++ src/Sylius/Bundle/ShippingBundle/Form/Type/Calculator/WeightRateConfigurationType.php
@@ -51,7 +51,7 @@ class WeightRateConfigurationType extends AbstractType
                 'label' => 'sylius.form.shipping_calculator.weight_rate_configuration.amount',
                 'constraints' => array(
                     new NotBlank(),
-                    new Type(array('type' => 'numeric')),
+                    new Type(array('type' => 'integer')),
                 )
             ))
             ->add('division', 'number', array(
